blog_django_graphql_react_relay/requirements/base.txt
Django==1.10.4

# Configuration
django-environ==0.4.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

djangorestframework==3.6.3

# GraphQL
graphene-django==1.3

requests==2.11.1
